+++ src/components/VueCommand.vue
@@ -4,17 +4,21 @@
     :class="{
       'vue-command': !invert,
       'vue-command--invert': invert
-    }">
+    }"
+    :style="font ? { 'font-family': font } : {}">
     <!-- Bar -->
-    <slot name="bar">
+    <slot
+      v-if="!hideBar"
+      name="bar">
       <div
-        v-if="!hideBar"
         :class="{
           'vue-command__bar': !invert,
           'vue-command__bar--invert': invert
         }">
         <div>
-          <slot name="buttons">
+          <slot
+            v-if="!hideButtons"
+            name="buttons">
             <span
               :class="{
                 'vue-command__bar__button': !invert,
@@ -42,16 +46,17 @@
           </slot>
         </div>
         <div>
-          <slot name="title">
+          <slot
+            v-if="!hideTitle"
+            name="title">
             <span
-              v-if="!hideTitle"
               :class="{
                 'vue-command__bar__title': !invert,
                 'vue-command__bar__title--invert': invert
               }">{{ title }}</span>
           </slot>
         </div>
-        <div />
+        <div>&#8203;</div>
       </div>
     </slot>
 
@@ -73,7 +78,6 @@
           'vue-command__history__entry--invert': invert,
           'vue-command__history__entry--fullscreen': shouldBeFullscreen(index),
           'vue-command__history__entry--fullscreen--invert': and(
-            invert, shouldBeFullscreen(index)
           )
         }">
         <component
@@ -103,7 +107,8 @@ import {
   computed,
   onMounted,
   nextTick,
-  getCurrentInstance
+  getCurrentInstance,
+  useSlots
 } from 'vue'
 import {
   createCommandNotFound,
@@ -133,6 +138,8 @@ import {
   debounce
 } from 'lodash'
 
+const slots = useSlots()
+
 const props = defineProps({
   commands: {
     default: () => ({}),
@@ -159,6 +166,12 @@ const props = defineProps({
     type: Array
   },
 
+  font: {
+    default: '',
+    required: false,
+    type: String
+  },
+
   helpText: {
     default: null,
     required: false,
@@ -172,6 +185,12 @@ const props = defineProps({
   },
 
   hideBar: {
+    default: false,
+    required: false,
+    type: Boolean
+  },
+
+  hideButtons: {
     default: false,
     required: false,
     type: Boolean
@@ -206,6 +225,14 @@ const props = defineProps({
     default: false,
     required: false,
     type: Boolean
+  },
+
+  // An interpreter allows to execute arbitrary code after a query has been
+  // dispatched
+  interpreter: {
+    default: null,
+    required: false,
+    type: Function
   },
 
   isFullscreen: {
@@ -287,6 +314,7 @@ const signals = reactive(newEventBus())
 const terminal = computed(() => ({
   cursorPosition: local.cursorPosition,
   dispatchedQueries: local.dispatchedQueries,
+  font: props.font,
   history: local.history,
   historyPosition: local.historyPosition,
   invert: props.invert,
@@ -352,6 +380,12 @@ const appendToHistory = (...components) => {
 // Parses the query, looks for a user given command and appends the resulting
 // component to the history
 const dispatch = async query => {
+  // Call optional interpreter to execute arbitrary code
+  if (isFunction(props.interpreter)) {
+    props.interpreter(query)
+    return
+  }
+
   // An empty query is an empty string
   if (isEmpty(query)) {
     appendToHistory(createQuery())
@@ -381,7 +415,7 @@ const dispatch = async query => {
   // instantly to history
   // TODO Find a better way to find out the name
   if (eq(get(command, '__name'), 'VueCommandQuery')) {
-    appendToHistory(command)
+    exit()
     return
   }
 
@@ -403,9 +437,9 @@ const dispatch = async query => {
   })
   appendToHistory(markRaw(component))
 }
-// Tear down component and execute final steps
+// Tear down component, execute final steps and return a new query
 const exit = () => {
-  // TODO Does order matter?
+  // TODO: Does order matter?
   appendToHistory(createQuery())
   setCursorPosition(0)
   setFullscreen(false)
@@ -521,13 +555,6 @@ onMounted(() => {
   vueCommandHistoryRef.value.addEventListener('scroll', eventListener)
 
   // Scroll to bottom if resized and scrolled to bottom before
-  const resizeObsever = new ResizeObserver(() => {
-    isTriggeredByResize = true
-    if (scrolledToBottom != null) {
-      // This fires scroll
-      vueCommandHistoryRef.value.scrollTop = vueCommandHistoryRef.value.scrollHeight
-    }
-    isTriggeredByResize = false
   })
   for (const vueCommandHistoryEntry of vueCommandHistoryRef.value.children) {
     resizeObsever.observe(vueCommandHistoryEntry)
@@ -552,7 +579,6 @@ provide('helpText', props.helpText)
 provide('helpTimeout', props.helpTimeout)
 provide('hidePrompt', props.hidePrompt)
 provide('incrementHistory', incrementHistory)
-provide('invert', props.invert)
 provide('optionsResolver', props.optionsResolver)
 provide('parser', props.parser)
 provide('programs', programs)
@@ -560,9 +586,10 @@ provide('sendSignal', sendSignal)
 provide('setCursorPosition', setCursorPosition)
 provide('setFullscreen', setFullscreen)
 provide('setHistoryPosition', setHistoryPosition)
+provide('setQuery', setQuery)
 provide('showHelp', props.showHelp)
 provide('signals', signals)
-provide('setQuery', setQuery)
+provide('slots', slots)
 provide('terminal', terminal)
 
 defineExpose({
@@ -606,20 +633,21 @@ defineExpose({
     }
   }
 
-  overflow-y: hidden;
   overflow-x: hidden;
+  overflow-y: hidden;
 
   .vue-command__bar,
   .vue-command__bar--invert {
     @include clearfix();
-    position: inherit;
+
+    display: flex;
+    font-family: -apple-system, BlinkMacSystemFont, sans-serif;
+    justify-content: space-between;
+    padding-bottom: 10px;
     padding-left: 10px;
     padding-right: 10px;
     padding-top: 10px;
-    padding-bottom: 10px;
-    display: flex;
-    justify-content: space-between;
-    font-family: -apple-system, BlinkMacSystemFont, sans-serif;
+    position: inherit;
   }
 
   .vue-command__bar__button,
@@ -641,32 +669,32 @@ defineExpose({
 
   .vue-command__history--invert,
   .vue-command__history {
-    overflow: auto;
-    word-break: break-all;
     display: block;
-    padding: 12px 12px 12px 12px;
-    margin: 0;
-    white-space: pre-line;
-    line-height: 1.33;
     height: 100%;
+    line-height: 1.33;
+    margin: 0;
+    overflow: auto;
+    padding: 12px 12px 12px 12px;
+    white-space: pre-line;
+    word-break: break-all;
 
     /* Provide reasonable default values */
     ul {
+      list-style-type: none;
       margin: 0;
       padding: 0;
-      list-style-type: none;
     }
 
     input,
     textarea {
       background: none;
       border: none;
-      outline: none;
       flex: 1;
       font-size: 1rem;
-      width: 100%;
-      resize: none;
+      outline: none;
       overflow: hidden;
+      resize: none;
+      width: 100%;
     }
   }